IAR - build and flash part of the IARWorkbench class
